HACK: doc/usage: library: document creating multiple objects of same type
TODO: Drivers should rather support a binding in the constructor instead
of calling set_binding_map.

While this should not be an everyday task to do via the API, it serves
as a good example to explain how bindings work. This knowledge may be of
good use later in the "Environments" section too.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>